fix: edit unused test
