+++ src/main/java/nextstep/member/Member.java
@@ -8,6 +8,8 @@ public class Member {
     private String phone;
     private Role role;
 
+    public Member() {}
+
     public Member(Long id, String username, String password, String name, String phone, Role role) {
         this.id = id;
         this.username = username;
